♻️ refactor: move workspace provider redirect into WorkspaceSetting feature
